feat: command handler
